# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SWP391 Drug Prevention API",
    "version": "2.0.0",
    "description": "Clean API documentation for SWP391 Drug Prevention Platform - Only active endpoints\n\n**🔓 DEVELOPMENT TESTING:**\nTo test protected endpoints without authentication, add the header:\n`x-swagger-bypass: true`\n\nThis will bypass token verification and use a mock admin user for testing purposes."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "components": {
    "schemas": {
      "DefaultResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Bearer {token}'"
      }
    },
...
  },
  "paths": {
...
    "/api/contents/{id}": {
      "get": {
        "tags": [
          "Contents"
        ],
        "summary": "Get content by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Contents"
        ],
        "summary": "Update content",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Contents"
        ],
        "summary": "Delete content",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/content/{id}": {
      "get": {
        "tags": [
          "Content"
        ],
        "summary": "Get content by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Content retrieved"
          }
        }
      }
    },
...
  }
}